Surprise reefactor of expand accordions code to make functions totally reusable.
