+++ _assets/js/print.js
@@ -1,24 +1,33 @@
-const preparePageForPrint = function() {
-  const beforePrint = function() {
-    document.querySelectorAll("details").forEach(detail => {
-      detail.setAttribute("open", true);
+import { toggleButtonText } from './utils/expandAccordions/setToggleButtonState';
+
+const preparePageForPrint = function () {
+  // Grab the button that will toggle the accordion opening:
+  const openAccordionsButton = document.querySelector('#crt-page--expandaccordions');
+
+  const beforePrint = function () {
+    document.querySelectorAll('details').forEach((detail) => {
+      detail.setAttribute('open', true);
     });
   };
-  const afterPrint = function() {
-    document.querySelectorAll("details").forEach(detail => {
-      detail.removeAttribute("open");
+  const afterPrint = function () {
+    if (openAccordionsButton) {
+      toggleButtonText(openAccordionsButton);
+    }
+
+    document.querySelectorAll('details').forEach((detail) => {
+      detail.removeAttribute('open');
     });
   };
 
   if (window.matchMedia) {
-    const mediaQueryList = window.matchMedia("print");
-      if (mediaQueryList.matches) {
-        beforePrint();
-      } else {
-        afterPrint();
-      }
-    };
-  
+    const mediaQueryList = window.matchMedia('print');
+    if (mediaQueryList.matches) {
+      beforePrint();
+    } else {
+      afterPrint();
+    }
+  }
+
   window.onbeforeprint = beforePrint;
   window.onafterprint = afterPrint;
 };
