+++ _assets/js/utils/expandAccordions/setDetailsElementState.js
@@ -1,0 +1,12 @@
+// Change the state of the details elements:
+const openDetails = (detail) => {
+    detail.setAttribute('open', 'open');
+    detail.setAttribute('data-detail-open', 'true');
+  }
+  
+  const closeDetails = (detail) => {
+    detail.removeAttribute('open');
+    detail.setAttribute('data-detail-open', 'false');
+  }
+
+export {openDetails, closeDetails};
